fix 32-bit build
git-svn-id: http://data-race-test.googlecode.com/svn/trunk@3001 93a3164d-3a44-0410-8ae5-4baa038b0729

+++ tsan/thread_sanitizer.cc
@@ -7132,7 +7132,7 @@ one_call:
     if (has_expensive_flags) {
       thr->stats.access_to_first_1g += (addr >> 30) == 0;
       thr->stats.access_to_first_2g += (addr >> 31) == 0;
-      thr->stats.access_to_first_4g += (addr >> 32) == 0;
+      thr->stats.access_to_first_4g += ((uint64_t)addr >> 32) == 0;
     }
 
     int locked_access_case = 0;
